plugins/renderers/image_renderer/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "image-renderer"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = ["pillow>=12.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."vesselframe.plugins.renderer"]
image_renderer = "image_renderer:make_plugin"
